Webapp update
setup for WebApp to let users interact with the model.

This was an update containing an attempt to use a python script in a web app. However, this does not seem to have a lot of support or functionality. So I am taking a new angle of applying a trained model into tensorflow.js which seems to have a lot more support